refactor: impact metrics modal - label filtering and UX
