+++ frontend/src/component/impact-metrics/ChartConfigModal/ImpactMetricsControls/SeriesSelector/SeriesSelector.tsx
@@ -47,6 +47,7 @@ export const SeriesSelector: FC<SeriesSelectorProps> = ({
                 placeholder='Search for a metric…'
                 variant='outlined'
                 size='small'
+                required
             />
         )}
         noOptionsText='No metrics available'
